feat: add eBPF observability stack with Beyla, Grafana, Prometheus & Tempo
- Add Grafana Beyla DaemonSet (v3.33.0) for zero-code eBPF instrumentation
    across all polyglot microservices (Python, Go, Rust, Node, Java)
- Add Prometheus (v3.14.0) with static scrape target for Beyla metrics
- Add Tempo (v3.0.3) for distributed tracing with OTLP ingest
- Add Grafana (v13.0.7-ubuntu-slim) with pre-provisioned APM dashboard

- Multi-row APM dashboard with $service/$route template variables
- Executive KPIs: RPS, error rate, p95 latency, active VUs
- Separated p95 and p50 latency panels per microservice
- Per-microservice HTTP status code distribution (full-width)
- Database & Cache row: PostgreSQL query throughput + p95, Valkey ops
- Downstream RPC client call rate panel

- Rust: 1.97.1 → 1.98.0 (Dockerfile, Dockerfile.migrations, image.py)
- Beyla: latest → 3.33.0
- Grafana: 11.5.2 → 13.0.7-ubuntu-slim
- Prometheus: v3.2.1 → v3.14.0
- Tempo: 2.6.1 → 3.0.3

- Add k6 checkout load test with 3 scenarios:
    - Standard checkout (80%), high-value multi-item (10%), fault injection (10%)
- Add k8s/k6-job.yaml for in-cluster Job execution
- Add smoke-test.js for quick validation
- Load profile: 30VU warm-up → 80VU sustained → 150VU peak over 5 min

- Add scripts/port-forward.sh for Grafana/Tempo/Prometheus/Gateway access
- Add scripts/run-k6.sh supporting local and k8s execution modes
- Add k8s/kustomization.yaml
- Fix payment serialization: amount field float (not Decimal) for Go Gin compat
- Update k8s/README.md with correct component versions and architecture notes

+++ gateway-py/src/gateway_py/orders/schemas.py
@@ -25,7 +25,7 @@ class PaymentRequest(BaseModel):
 
     order_id: UUID
     customer_id: str
-    amount: Decimal = Field(..., gt=0)
+    amount: float = Field(..., gt=0)
 
 
 class InventoryReserveItem(BaseModel):
